fix: 배고픔 게이지 마이너스 상수 수정

+++ src/main/java/com/roomie/server/domain/roomie/application/RoomieService.java
@@ -56,9 +56,9 @@ public class RoomieService {
 
             Duration timeDifference = Duration.between(lastFeedTime, now);
 
-            Long differenceToSeconds = timeDifference.getSeconds();
+            long differenceToSeconds = timeDifference.getSeconds();
 
-            Long hungerMinusAmount = differenceToSeconds / StaticValue.HUNGER_MINUS_SECOND_STRIDE;
+            Double hungerMinusAmount = (double) differenceToSeconds / StaticValue.HUNGER_MINUS_SECOND_STRIDE;
 
             double roomieHungerGage = roomie.getHungerGage() - hungerMinusAmount;
